+++ apps/trading-e2e/src/integration/market-info.cy.ts
@@ -1,4 +1,4 @@
-import { Schema, MarketTradingModeMapping } from '@vegaprotocol/types';
+import { MarketTradingModeMapping } from '@vegaprotocol/types';
 
 const marketInfoBtn = 'Info';
 const row = 'key-value-table-row';
@@ -209,43 +209,4 @@ describe('market info is displayed', { tags: '@smoke' }, () => {
         cy.get('dd').should('contain.text', value);
       });
   }
-});
-
-describe('market states', { tags: '@smoke' }, function () {
-  //7002-SORD-062
-  //7002-SORD-063
-  //7002-SORD-066
-
-  const states = [
-    Schema.MarketState.STATE_REJECTED,
-    Schema.MarketState.STATE_CANCELLED,
-    Schema.MarketState.STATE_CLOSED,
-    Schema.MarketState.STATE_SETTLED,
-    Schema.MarketState.STATE_TRADING_TERMINATED,
-  ];
-
-  states.forEach((marketState) => {
-    describe(marketState, function () {
-      beforeEach(function () {
-        cy.mockTradingPage(marketState);
-        cy.mockGQLSubscription();
-        cy.visit('/#/markets/market-0');
-        cy.wait('@Market');
-        cy.connectVegaWallet();
-      });
-      it.skip('must display correct market state');
-      //7002-/SORD-/061 no state displayed
-      it('must display that market is not accepting orders', function () {
-        cy.getByTestId('place-order').click();
-        cy.getByTestId('dealticket-error-message-summary').should(
-          'have.text',
-          `This market is ${marketState
-            .split('_')
-            .pop()
-            ?.toLowerCase()} and not accepting orders`
-        );
-        cy.getByTestId('place-order').should('be.disabled');
-      });
-    });
-  });
 });
